requirements.txt
streamlit
numpy
joblib
scikit-learn
rdkit-pypi